test(#zimic): split large test files by feature
