+++ packages/zimic/src/interceptor/http/interceptor/__tests__/shared/methods/patch.ts
@@ -14,7 +14,7 @@ import { expectFetchError } from '@tests/utils/fetch';
 import { usingHttpInterceptor } from '@tests/utils/interceptors';
 
 import { HttpInterceptorOptions } from '../../../types/options';
-import { RuntimeSharedHttpInterceptorTestsOptions } from '../interceptorTests';
+import { RuntimeSharedHttpInterceptorTestsOptions } from '../types';
 
 export async function declarePatchHttpInterceptorTests(options: RuntimeSharedHttpInterceptorTestsOptions) {
   const { getBaseURL, getWorker, getInterceptorOptions } = options;
